Screen('ColorRange'): Don't force color clamping on if unspecified.
Rather dumb behaviour was to assume clampcolors = 1 whenever no
optional clampcolors setting was given, instead of the less insane
option of just leaving the setting as it is. This caused an innocent
'ColorRange' query to change the clampcolors setting as a side effect,
e.g., each call to WhiteIndex() would execute a white =
Screen('ColorRange', window); query and thereby tamper with the
clampcolors setting. Hilarity ensues!

Switch to sane behaviour of leaving the setting alone if it isn't
specified by usercode.

Screen for 64-Bit Matlab on OSX rebuilt & tested, fixes #88.