PsychKinectCore: Fix wrong matrix dimensions of 'GetImage'
PsychKinectCore('GetImage') returned a matrix with flipped height and
width dimensions, requiring awkward reshape() operations to make any
use of the returned matrix. Now fixed. Also added a description to the
online help text on how to convert the returned image matrix into
something Matlab/Octave's image processing toolbox / imshow() /
imwrite() et al. understand. Some help texts improved.
Clarified that the driver currently only suppports XBOX-360 Kinect, but
not yet the "Kinect for Windows".

PsychKinectCore for 32/64-Bit Matlab on Windows and 64-Bit
Octave/Matlab on OSX rebuilt. Rebuild for Linux pending. We won't
rebuild or fix the driver for 32-Bit Matlab on OSX anymore, as 32-Bit
PTB for OSX is deprecated and this is the kind of exotic and work
intense maintenance stuff we won't do anymore for a dying platform.

Also cleanup to the demos.