Screen/Windows: Rebuilt + Compiler (warning) fixes.
Rebuilt for testing. Fixed a compile error in new Gstreamer vidcap
code. Fixed various compiler warnings, so that it now compiles almost
cleanly, except for some deprecation warnings wrt. to GStreamer which
we have to live with for the time being on all platforms.

Screen for 32/64-Bit Matlab on Windows rebuilt.