Screen: Implement PsychCleanupDisplayGlue() on all platforms.
This function, formerly Linux only, now also exists on OSX and Windows.
It is called during Screen shutdown to release the display glue - the
stuff that happens in PsychScreenGlue. The routine shall cleanup
whatever display and graphics related resources are held while the
Screen() file is loaded.

On Linux it releases XInput and X11 display connections -- basically
disconnects from the X-Server.

On OSX it disconnects from the PsychtoolboxKernelDriver and releases
CGDisplayReconfigurationCallbacks. This was formerly done in a hacky
way.

On Windows it releases display system resources.

-> Much cleaner solution, no more hacks here.