Screen/GStreamer: Add setup of multi-threaded video decoding method.
If multi-threaded video playback is requested, check if the
"thread-types" property is supported on the selected video codec. If
so, set it so that the codec uses both slice-threading and
frame-threading for parallel decoding.

By default, the GStreamer-SDK's codecs and current upstream master only
use slice-threading for decode, because frame-threading can cause
visual artifacts with some movie formats and has higher latency, which
can be bad for video teleconferencing applications and such. However,
we only do latency insensitive playback of formats we know that work
with frame-decoding (opt-in of users), and frame-decoding has much
higher performance/scalability. Therefore we absolutely want
frame-threading in addition to slice-threading and need to set the
property to a proper non-default value, as the default would be
slice-threading only.

The property itself is not yet implemented in GStreamer at time of this
commit. This is tbd by me soon in upstream GStreamer.

See
http://cgit.freedesktop.org/gstreamer/gst-ffmpeg/tree/ext/ffmpeg/gstffmp
egviddec.c?h=0.10#n519 for a reference to the code that needs to be
changed in the gst-ffmpeg plugins. See
https://bugs.freedesktop.org/show_bug.cgi?id=56206 for the discussion
on this gstreamer feature request.

Screen for 64-Bit Matlab on Windows rebuilt.