Screen/Linux: Add standard NETWM property to auto-disable compositors.
The KDE/KWIN method to disable desktop compositors has been
standardized as a NETWM standard. The property
_NET_WM_BYPASS_COMPOSITOR if set on a fullscreen onscreen window will
ask the compositor to auto-disable as long as the window exists. Set
this property on our windows additionally to the "old" KDE/KWIN
property, so we cover the past and the future.

All compositors are supposed to honor this property in the future. The
first one to implement it in its master branch is "Mutter" the
compositor for GNOME-3, as of 18th December 2012.