Screen: Allow check if desktop composition is active on all platforms.
So far we only had heuristics for Windows Vista/7/8. This refactors
code a bit to implement PsychOSIsDWMEnabled(screenNumber) for all
platforms. OSX should just work but is basically pointless, implemented
anyway.

Linux is not clear, we may need to disable this again for the time
being - test pending. The problem is that the detection code
(supposedly) can detect if composition on a specifix screen is active
or not, but that info may not be useful because we want to know if it
is active for our window, not for the whole screen on which our window
resides. According to mailinglists (test pending), KDE/KWIN suspends
desktop composition for a whole screen if asked by us, so screen state
== windows state -> all good. But other WM's, e.g., apparently
Unity/Compiz do not suspend composition for whole screen, they just
suspend it for our fullscreen windows via unredirection, so while our
window may not suffer composition, the method still returns
"composition enabled", as it wasn't globally disabled for the whole
x-screen. In that case the new check would only produce false alarms at
least on Unity/Compiz - thereby would be harmful instead of helpful.

Some testing required. Screen for 32-Bit Octave on Linux rebuilt. Will
see what we do...