Screen/Linux: Make WM hints for disable of compositors actually work.
At least KDE/KWin didn't honor our properties to disable desktop
composition. The culprit seems to be our window being override_redirect
= true. Removing that, and setting the window fullscreen via
_NET_WM_STATE_FULLSCREEN property, makes it work. KWin nicely disables
desktop composition as soon as our opaque fullscreen window appears :-).

However, not all is shiny and rosy. On multi-display setups, the window
manager ignores the (x,y) position of the top-left corner of our window
and  picks one display monitor and positions the window so it goes
fullscreen on that monitor - it doesn't cover the whole x11 screen,
which is bad, as only then we get maximum performance and precision due
to page flipping.

So what to do? First we have a new conservevram setting that allows to
go back to the old override_redirect strategy which served us well for
many years without any problems, but doesn't allow auto-disable of
compositor.

Then the solution (hack?) for now is to let the WM have its way and
misplace the window, but after the window has been mapped, we set
the override_redirect = true to lock the WM out of future window
management and then move the window to its proper location, without
the WM noticing or complaining.

Net result is we get proper windows under KDE/KWin, but also the
benefit of automatically disabled compositing, as long as one of
our standard fullscreen onscreen windows is active.

This commit contains a working Screen mex file for 64-Bit Octave
on Linux. Successfully tested with KDE/KWin.