Screen/Linux: Apply PsychOSIsDWMEnabled() only for debugging.
...not for real reporting, as the check only works reliably on KDE/KWin
and produces endless false positives on other compositors, e.g.,
GNOME-Shell and Compiz.

Reason is that KWin has a different standby behaviour. It actually
disables compositing for a whole x-screen on request and releases all
resources, whereas other compositors still composite other windows on
the x-screen, they only unredirect our onscreen window, so our test for
global disable on a x-screen is too coarse/global to work on non-KDE.

Anyway leave the check in at high verbosity levels > 10 for testing and
debugging purposes, not as a "end user feature".