VBLSyncTest: Add workaround for Datapixx + Octave bug.
A strange bug. If Datapixx timestamping is used under Octave,
then Screen('Flip') only returns its first tvbl argument, but
not the other ones. Needs investigation, but is too close to
release, so we plaster over it with this workaround.

Causes fake-discrepancy of 1 vblank duration ~ 4% or 0.5 msecs
on a typical display at typical refresh and resolution.