Screen/OSX: Try to handle MBP Retina scaled mode "best for retina".
Contrary to what one would assume, the default display mode on
MacBookPro's with Retina displays called "Best for retina" doesn't
operate at full native panel resolution (e.g., 2880 x 1800) but only at
quarter resolution (e.g., 1440 x 900) and rendered content is then
upscaled to full panel resolution. This means that the backend operates
at twice the vertical resolution of the frontend and vbl_startline is
not 900 but 2 * 900, as is vbl_endline. Handle this by a slight
modification of our code for detecting double-scan video modes
originally developed for Linux. We (ab)use the same code for handling
Retina mode on OSX. As we don't know if the timing/timestamping is
actually correct, warn user about scaled retina mode and timing
implications. The problem is that we don't know how the scaling is
implemented and if it introduces some sort of implicit triple-buffering
which would break all our timestamping algorithms. The only way to find
out would be for a user to use a photo-diode and lots of masochism and
experience to find it out.

Apparently it is not possible to select native resolution display with
Apple's display setup panel, but some 3rd party tools can configure it:

<http://apple.stackexchange.com/questions/71636/can-i-use-a-retina-displ
ay-at-actual-pixel-size?rq=1>

<http://apple.stackexchange.com/questions/59964/macbook-pro-retina-setti
ng-resolution-to-2880-restored-to-1440-after-restart-sys?rq=1>

<http://www.anandtech.com/show/5996/how-the-retina-display-macbook-pro-h
andles-scaling>

Another fun problem not addressed here is that OSX native beamposition
queries for NVidia cards at least for Retina panels seem to be
completely broken. Unclear if it is only Retina or OSX 10.8 + NVidia in
general. Anyway, the only way to get any (potentially meaningful)
timestamping is to load the PsychtoolboxKernelDriver and use
'conserveVRAM' option 2^16 to disable native beamposition queries so
our own code is used. Screen can't detect the need for this yet.

-> Screen for 64-Bit Octave and 64-Bit Matlab on OSX rebuilt.
-> Won't upgrade 32-Bit Screen for this release, as 32-Bit PTB for OSX
is deprecated.
-> help BeampositionQueries updated with troubleshooting tips.