PsychImaging: Add support for xyY -> XYZ and XYZ->RGB color space conversion.
Add support for imaging pipeline implementation of the
SensorToPrimary() and xyYToXYZ() M-Functions.

This allows to draw into the framebuffer in xyY chromacity+luminance
format and auto-convert to RGB framebuffer values at Screen('Flip')
time via imaging pipeline. It also generally allows to convert
XYZ->RGB, which is the basis for support for other possible input color
spaces.

RenderDemo.m now demonstrates all methods to display xyY format, given
a 'cal'ibration struct and gamma table from display calibration.

This is based on the description given by David Brainard in
Psychtoolbox forum message #15092.

The code seems to work correctly and with high precision, but is only
very lightly and manually tested. A unit test is needed for safety.
This is a backup commit to document the basic working state.