GPUTypeFromToGL: Handle 1-D, 2-D, singleton dimension and complex input matrices.
1D or 2D matrices weren't handled correctly. Fix this by reshape()'ing
them into 3D matrices with singleton dimensions. 1D -> 2D
single-channel luminance texture with a single row. 2D -> 2D luminance
texture.
3D matrices of more than 4 channels must be rejected, as they aren't
convertable to 2D textures. Possible future improvements would be using
3D textures, texture arrays or PBO's in such cases, also for matrices
with four or more dimensions.
3 layer matrices, which would convert into RGB textures which are not
supported by CUDA-OpenGL interop, are extended to 4 layers, the 4th
layer filled with 1.0 for a "neutral" alpha channel in the resulting
RGBA texture. Such 3-layer matrices trigger a upcast warning, so user
is not left confused about what magically happened.
Complex matrices which store complex numbers: Currently a performance
warning is given and only the real part of the complex number is
stored, the imaginary part is thrown away. For the future we may add an
option to allow storing the imaginary part in the alpha channel of a LA
texture, or the RG channel pair of a RGBA texture, or similar, to
retain this info. FFT's produce complex numbers as output and i could
imagine good use cases for FFT's in usage scenarios of Psychtoolbox, so
being able to represent at least 2D matrices of complex numbers like
from a 2D FFT could come in handy.

Not sure what to do about outputs. Currently a single layer luminance
texture is returned as a 3D matrix with the 1st dimension being a
singleton dimension. We could squeeze() that away, but i'm not sure if
it isn't better to leave that to the usercode to provide a bit more
control? For now we don't squeeze them.

This makes the routine working with GPUmat's fft2() and ifft2() 2D-FFT
functions.