Screen/Linux: Remove nouveau boilerplate warning.
Nouveau, as of version 1.0.0, released 18th June 2012, is
no longer a hazard for visual stimulus timing. It should
provide accurate/robust stimulus onset timing. Stimulus
onset timestamping should be reliable and robust. Accuracy
of onset timestamps is still not as great as it could be
due to lack of high precision support in the nouveau-kms
kernel driver as of Linux 3.9, februrary 2013. However,
accuracy is comparable to what other vision science toolkits
would achieve and good enough for light to medium precision
needs, about 1 msec. Usercode can switch to high precision
by using Screen('Preference','VBLTimestampingmode', 1); if
needed. This (plus the usual amount of tinkering with display
mappings on non-standard display setups) would use the
high-precision beamposition timestamping as with the
proprietary NVidia binary blob.

Anyway, do away with the warning boilerplate, it is no
longer justified.