Screen/Linux: Initial check-in of Waffle display backend.
PsychWindowGlueWaffle.c implements a Linux display glue
layer based on Chad Versace's Waffle rendering/display
backend selection library:

http://cgit.freedesktop.org/~chadversary/waffle

It will be used to run the Linux PTB with other display
backends than X11+GLX and other rendering backends than
OpenGL-2 desktop core. This is mostly useful for gently
testing out the waters with Wayland, and for porting to
embedded/mobile devices like Android tablets/phones and
things like Rasperry Pi et al.

It is unlikely that a pure Waffle backend, at least in
unmodified form, will be a good permanent fit. Likely
either Waffle needs to be customized/extended for our
use, or we need to implement our own backends. But as
a cheap and relatively pain-free starter it should
hopefully be good enough.

This is a backup commit: The file is just a copy of
our regular X11+GLX display glue, with indent run over
it to cleanup the white-space and formatting mess that
accumulated over the years.