Added two utility routines from BrainardLabToolbox
GetSVNInfo -- returns SVN version number of directory or file.  Useful
for recording state of software at time an experiment or calculation
was done.

BasicToneMapCalFormat -- very simple tone mapping.

I moved both of these to PTB because a) they might be generally useful
and b) so that RenderToolbox3 users don't need to get the
BrainardLabToolbox.

DHB