Rearrangement of some radiometric routines, documenting and improving tests.
The test programs are now in pretty good shape.  I don't get complete
agreement with the calculations made by Francois' Delori's spreadsheet.
 It may be that the 2000 version of the standard differs enough from
the 2007 standard that the divergence is not actually error.  Am trying
to consult with Delori and other experts to track this down.

Added cautionary notes throughout the MPE computation routines, and in
particular that we recommend checking calculations independently before
shining bright lights into subjects eyes.

Will continue to chip away at this stuff, and to try to rationalize the
routines somewhat.