Screen: Enable OpenGL-ES FBO's if OES_framebuffer_object is supported.
Detect extension, and enable relevant parts of imaging pipeline, if
possible.

FBO's can be used for fast offscreen window support, render-to-texture
via on-the-fly conversion of textures into offscreen windows on first
draw, and to enable at least those parts of the imaging pipeline which
can do without shaders, as we currently only support the fixed-function-only
OpenGL-ES 1.x.