Screen/Waffle: Temporarily disable eglSwapInterval().
The proprietary desktop OpenGL drivers from NVidia currently
don't support EGL, so we couldn't test OpenGL-ES support on
those if we had eglSwapInterval() as undefined symbol in our code.

PsychOSSetVBLSyncLevel() will be a no-op on EGL backends, but
testing on desktop Linux + NVidia binary blob should be possible.

No big deal, as the default swapinterval of EGL is 1 == what we want
99.9% of the time anyway.