Screen/Waffle: Make moglcore functional.
Screen now exports the finally chosen display backend and rendering
api to the host runtime by setting the same environment variables
that can be used to override its choice of backend/api.

moglcore on Linux checks the PSYCH_USE_DISPLAY_BACKEND variable
and if it exists (== Waffle backend instead of classic X11/GLX backend
in use), what display backend is in use. If the backend isn't GLX,
then it skips glewInit(), which would crash on a non GLX setup, and
instead relies on the fact that the special glewInit() inside Screen()
will have already setup all GL entry points and symbols, exposing them
to the global address space - and thereby moglcore - due to the Screen
re-dlopen() procedure. On other os'es or with glx backend or classic
backend, moglcore performs glewInit() as usual.

-> This makes moglcore functional. However, we basically don't have
   any interesting demos for testing, as almost all of our code
   uses immediate mode glBegin()/glEnd() rendering, which is not
   supported on OpenGL-ES rendering apis.

=> Screen.mex and moglcore.mex for 64-Bit Octave on Linux rebuilt.