BitsPlusImagingPipelineTest and BitsPlusIdentityClutTest: Add Bits# support.
Allow to run Bits# hardware assisted diagnostic routines and
auto-fixes, similar to the ones already available for VPixx
DataPixx, ViewPixx and ProPixx.

Also allow specification of an optional 'windowrect' for the
onscreen test window. This only makes sense on Linux for quick
and dirty testing, not for production use on any os.