Screen for 64-Bit OSX: Fix linker failure on 10.5 Leopard due to incompatibility.
10.5 Leopard 64-Bit does not support the Cocoa class NSRunningApplication,
causing runtime linker failure ("Invalid MEX file") when trying to use it
on 64-Bit Matlab or Octave on 10.5.

Fix this by dynamically detecting and binding the class dynamically and
no-op'ing if it is unsupported on 10.5. No-Op will cause SetFocusWindow()
to silently fail on 10.5, which means that window keyboard input focus
isn't restored for the Octave or matlab -nojvm console window at fullscreen
window close on 10.5, so user has to manually click the window.

This is a minor inconvenience for an old and outdated os.

Screen for 64-Bit Octave/Matlab on OSX rebuilt and tested on 10.7.
Test by user for 10.5 pending.