BitsPlusIdentityClutTest: Work around a MS-Windows graphics bug.
Not clear yet if it is a bug in the Windows graphicd drivers, the
windows GDI text renderer, or our code, but with text alpha blending
enabled, we get off-by-one color index errors in text rendered to the
color overlay. This results in off-by-one color palette clut access and
weirdo color artifacts which suggest failure of the imaging chain, when
everything is actually fine.

Disabling text alpha blending fixes the problem on Windows.

Also disable the 'o' key on Windows, which works by setting the
blue channel of the gpu gamma table to zero - or actually doesn't
work on Windows, where such gamma tables are illegal and cause
error abort due to failure of 'Loadnormalizedgammatable. This is
now a OSX and Linux only feature.

Minor code cleanups.

-> All tests and demos now pass and work correctly with the Bits#
   on MS-Windows in all display modes.