Screen/ES: Handle ES implementations with only fixed point and line size 1.
If an OpenGL-ES implementation only supports a fixed point diameter or
line width of 1 pixel then skip the size checks and stick to a fixed
1 pixel size for Screen('DrawDots/DrawLine/DrawLines'). This is needed,
e.g., on Tegra-3 of the Google Nexus-7.

Normally we are not lenient if scripts try to select unsupported point/
line sizes, but to get anything going on embedded devices we are more
forgiving here. I guess researchers would note during testing of code
if a varying dot/line size stimulus would show up only with 1 pixel thin
lines or dots, unless the person is really incompetent.