Screen/Nexus-7: Use proper order of GL libraries.
There is no way we can fix the libGL before all else problem
of Octave as build for the Nexus-7, only workarounds. Our options
are:

1. File a bug with Debian/Ubuntu to build octave for embedded
   devices without OpenGL support.

2. Stopgap1: LD_PRELOAD=/usr/lib/libGLESv1_CM.so the ES-1 library
   before everything else at Octave startup, as long as 1. hasn't
   taken effect.

3. Stopgap2: Strip liboctinterpreter.so from all dependencies on
   libGLU and libGL, esentially the brute-force dot it yourself
   implementation of 1. This could be done by a bundled ptb
   setup script.

In order to allow 3. we need to provide the GL libs in an order
that works for us, as we still depend on GL to satisfy Screen.mex
dependencies and GLU (which depends on GL) for the tesselators
needed for Screen('FillPoly'). The GL lib is dead but that doesn't
matter, luckily.

Update build script to link in order -lGLESv1_CM -lGL -lGLU, so
everything needed is there, but GLES 1.1 takes precedence over
desktop GL for all gl entry points which we actually need in a
functional state.

Screen for ARM rebuilt - works.