Add prebuilt libwaffle.so's for Ubuntu+ARM+Nexus7
Cohorts/libwaffle/ contains some prebuilt waffle.so's
as building from source may be a bit tedious for early
adopters.

The _glexegl version has support for X11/GLX and X11/EGL,
as needed on the Nexus-7 with NVidia Tegra-3 binary drivers
on Ubuntu 13.04.

The _withgbm version has additional support for the GBM/EGL
backend, but this is not useful on my device. May be useful
on other embedded devices with FOSS graphics drivers where
GBM/DRM/KMS is supported.