Screen/Waffle: Fix OpenWindow races/hangs on Nexus.
Whenever we transition between waffle and x11 operations,
we need to flush and sync the x11 display connection via
call to XSync() if a x11 based display backend is active,
as we use classic xlib, but waffle uses xcb, and bad hangs
and races can happen if both are intermixed without proper
sync.

Now fullscreen windows work properly not only on first run
in a session, at least on xfce without compositor. Tests
with compositor and unity pending...