Screen/Linux: Select specially named drawtext plugin on ARM architecture.
On Linux, if running on a ARM architecture processor, select the
text plugin which is suffixed arm or arm64 before trying the
generic name.

This doesn't matter for NeuroDebian/Debian/Ubuntu packaged PTB's, as
there is only one drawtext plugin for the given machine architecture,
but it is important for our own upstream distribution, so we can separate
the intel arch files from the arm arch files.

Screen for 64-Bit Octave on Linux rebuilt and tested.