Screen: Fix roundoff misbehaviour for POT textures.
The way we computed the power-of-two size for a texture
when blitting it in the non-power-of-two GL_TEXTURE_2D
path was fragile. It used double instead of int types,
so was subject to potential roundoff errors.

No big deal on Intel or PowerPC with their high precision
fpu's, but a nice little blowup on ARM with its apparently
lower precision fpu.

Fix it. Test on ARM pending.