Screen/ES: Optimize classic dontclear=1 mode of Flip.
If framebuffer needs to be backed up to textures due to
lack of AUX buffers, try to use more optimal but optional
GL_BGRA_EXT storage format if OES extension is supported.

This gives 42% speedup with MouseTraceDemo2 on Nexus7
with Nvidia Tegra-3 gpu. However, in practice that's a
performance of 6 fps instead of 4 fps - still abysmal!

Proper solution: Use the imaging pipeline if supported - Fast!