Updated cal file routines to search one level of subdirs.
Added code that causes LoadCalFile to look for the specified
calibration file up to one level deep of subdirectories of the
calibration file directory, and SaveCalFile to behave correspondingly. 
I need this
so I can organize my calibration folder more sensibly.

This code is a bit hairy because it has tried to maintain backwards
compatibility all the way to a system which made sense under Mac OS9.
I tried to make these changes so that all old behaviors continue to
work.   It is possible I broke something however.  Keep an eye out.

We will test this in my lab for a while and try to fix up anything that
doesn't work.  We don't exercise all the old ways of calling these
routines, however.

DHB