Linux 32-Bit drawtext plugin rebuilt with fix for idiotic bug.
Because the -l linker specs were defined *before* the c code, no
dependencies on the required GL, fontconfig and freetype libraries were
encoded in the plugin! Worked by accident on 32-Bit octave on Linux,
but not on 32-Bit Matlab, because octave also links against freetype et
al., but Matlab doesn't.