Screen: Allow efficient audio volume change for GStreamer movie playback.
If the Screen('Playmovie', movie, rate, loop, volume); function is
called while a 'movie' is already playing, passing in a 'rate' and
'loop' flag that is identical to the current playback rate and loop
setting, ie., a redundant setting, then allow to change audio mute
state and playback volume setting 'volume' to be applied with very low
overhead.

This allows on-the-fly changes of audio movie playback volume or
(un)muting of audio during playback.