Polish initial GPGPU CUDA/GPUmat compute support.
A new PsychImaging() subfunction 'UseGPGPUCompute' must be called to
detect+select+initialize the GPGPU computing api before it can be used
by PTB. Currently only the GPUmat toolbox for NVidia CUDA computing
with Matlab is supported, but the PsychImaging et al. basic interface
should be sane enough to allow to extend this eventually to OpenCL and
octave.

Demos and help texts have been updated and cleaned up.
This is good enough for the initial release, but much work remains to
be done to make this nice and extend its useful functionality beyond a
bit of FFT based image filtering.