Change post-installer to use javaclasspath.txt instead of classpath.txt
The masters of good software engineering and backwards compatibility
have changed the name and location of the java static class path
definition file in R2013a (v8.1) and later.

No longer shall we write to classpath.txt, but instead to javaclasspath.txt,
which can be located in the Matlab startup folder, or the users preference
folder.

Adapt: On v8.1+ we try to find the javaclasspath.txt file. If we can't
we assume it is in the users local preference folder (which we create if
it doesn't already exist) and set the path and filename accordingly.

On earlier versions, we still use classpath.txt as in the past.

PsychtoolboxPostInstallRoutine and DownloadAdditionsForLinux updated.

-> This change is entirely untested due to lack of R2013a access.