Screen: Allow to create hidden/invisible onscreen windows.
If Screen('Preference','WindowShieldingLevel', -1); is set
before/during call to Screen('Openwindow') for a non-fullscreen
window with kPsychGUIWindow 'specialflags' set, then create
the onscreen window as usual, but do not show/map it, so it
stays invisible on the display.

This is useful if one wants to (ab)use Psychtoolbox as an
offscreen renderer. One can create such an onscreen window,
then an associated 'Offscreenwindow' and then render and
'GetImage' from that offscreen window. The offscreen window
can have arbitrary size and resolution (within the hardware
and vram memory limits of the gpu of course).

One should probably disable sync tests etc. as in
PsychDebugWindowConfiguration() to avoid sync errors
or lots of warning clutter for such a window.

Screen for 64-Bit Octave on Linux rebuilt and tested.