DAQ toolbox: Another attempt to make it less broken.
Instead of calling PsychHID('Devices') directly, all DAQ functions now
call the new function PsychHIDDAQS. This function queries all HID
devices via PsychHID('Devices'), and caches the device list internally
for faster lookup. More importantly, it preprocesses all HID-DAQ
devices as follows: Non MCC devices are ignored. MCC devices get their
productID remapped to a standardized product name string. This because
on MS-Windows, PsychHID can't return a sensible product name string,
but all DAQ toolbox functions are dependent on name string matching.
The productID is well defined across devices and operating systems, so
we use that to map  properly. Also, serialNumber is replaced by a
combination of serialNumber and locationID, again because depending on
OS and device vendor and type, either one of those two can be undefined
or unreliable or useless. Combining both into a sort of hash will
hopefully be at least somewhat robust to common case usage.

DaqDeviceIndex: Fix some bugs and make more robust. This one should be
able to identify devices properly, based on PsychHIDDAQS preprocessing.
It should be able to reliable count and disambiguate interfaces if
multiple device are connected. And fix some bugs. On Linux and Windows,
the interfaceID of a DAQ device is well defined, so finding the daq
index corresponding to interface #0 should be reliable. On OSX we don't
have interfaceID, so we still need to rely on the heuristic of trying
to find interface #0 based on the number of outputs. Given that that
number changes without rhyme and reason across versions of OSX, and
sometimes apparently even across reboots, OSX will remain the least
reliable OS for use with the DAQ toolbox, despite the DAQ toolbox
originally written for it. Yay Apple!

-> All this is untested so far, but given how broken this stuff is, it
can't get much worse.