Update some DataPixx mex files.
These are from the latest zip file on VPixx website. However there
functionality is inconsistent across files, e.g., some have
VesaStereWaveform and TouchPixx support, others don't.

However, at least the Matlab mex files for Windows and OSX seem to be
up to date. This is important especially for 32-Bit Matlab on OSX, as
we're going to cancel that ptb very soon, so we want to end it with an
up to date Datapixx mex file for 32-Bit Matlab on OSX.