Factor out Matlab Java classpath setup into PsychJavaTrouble(1);
Instead of having multiple copies of the setup code, implement it in
PsychJavaTrouble and execute it if PsychJavaTrouble(1); is called. This
is now used from the PsychtoolboxPostInstallRoutine during regular
installation and update/setup, and from DownloadAdditionsForLinux for
legacy install via outdated v3.0.9 Debian or Ubuntu Psychtoolbox
archives. In case of up to date packages or current NeuroDebian
packages, users should call PsychJavaTrouble(1) once from within Matlab
directly, in additions to PsychLinuxConfiguration.

Update online help text of PsychJavaTrouble to state of PTB 3.0.10+