Supress irritating warning
The MATLAB function str2num issues a warning if the string it is passed is longer than 63 characters.  I think this is because it first tries treating its argument as a MATLAB variable name, and then goes on to try it as a string.

I wrapped the call to str2num in code that turns off this warning, calls str2num, and then restores the warning state to what it was.  Since we always passing a string here, this should not cause any failures to warn when appropriate.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>