Scripting glue: Add workaround for octave mexErrMsgTxt("") bug.
Octave's interpreter doesn't abort script execution as it should
if an empty "" error string is passed to mexErrMsgTxt(), as we
did if Screen('Openwindow') failed. This is true as of current
Octave 3.6 and can cause some confusing error behaviour in some
cases.

Workaround by passing some generic error message if the error
string from caller is empty.