Screen/GStreamer: Make encoding option settings work with camerabin2 video recording.
camerabin2 video recording uses GstEncodingProfile's with loadable
preset settings to define selection and configuration of encoders and
muxers for video recording, so our manual codec option setup doesn't
work. Make it work by:

1. Creating and setting options on all codecs as with camerabin1 or
movie writing operation.
2. Saving codec settings as GstPreset's to filesystem.
3. Loading those presets in encoding profile setup.

Sort of works, except the x264enc encoder has a bug which causes it
to crash if trying to save presets, so it is black-listed and presets
for it need to be provided in a premade file that one needs to copy
to the $HOME/.gstreamer-0.10/presets/ folder. The file is stored in
PsychVideoCapture/GstX264Enc.prs and hard-coded to use fastest encoding
strategy.

However, turned out H264 default setup wasn't the bottleneck. It is
somehow related to audio recording. Video recording without audio
works with reasonable performance, but if audio recording is enabled
performance goes down the drain. Reason completely unknown, so far
tested on OSX 10.7.5.